issue config new expo version
