+++ apps/mobile/package.json
@@ -2,6 +2,7 @@
   "name": "mobile",
   "version": "1.0.0",
   "main": "index.js",
+  "packageManager": "pnpm@8.15.9",
   "scripts": {
     "dev": "expo start",
     "build:web": "npx expo export --platform web",
@@ -42,7 +43,7 @@
     "@dynamic-labs/react-native-extension": "4.0.0-alpha.8",
     "@dynamic-labs/utils": "^3.3.0",
     "@dynamic-labs/viem-extension": "4.0.0-alpha.8",
-    "@expo/metro-runtime": "~3.2.1",
+    "@expo/metro-runtime": "~3.2.3",
     "@getalby/bitcoin-connect-react": "^3.5.3",
     "@getalby/lightning-tools": "^5.0.3",
     "@getalby/sdk": "^3.6.1",
@@ -88,22 +89,22 @@
     "echarts": "^5.6.0",
     "echarts-for-react": "^3.0.3",
     "events": "^3.3.0",
-    "expo": "latest",
-    "expo-application": "latest",
-    "expo-auth-session": "latest",
-    "expo-av": "latest",
-    "expo-camera": "latest",
-    "expo-checkbox": "latest",
-    "expo-clipboard": "latest",
-    "expo-constants": "latest",
-    "expo-crypto": "latest",
-    "expo-device": "latest",
-    "expo-font": "latest",
-    "expo-image-picker": "~15.0.7",
+    "expo": "53.0.7",
+    "expo-application": "^5.9.1",
+    "expo-auth-session": "^5.5.2",
+    "expo-av": "^14.0.7",
+    "expo-camera": "^15.0.16",
+    "expo-checkbox": "^3.0.0",
+    "expo-clipboard": "^6.0.3",
+    "expo-constants": "^16.0.2",
+    "expo-crypto": "^13.0.2",
+    "expo-device": "^6.0.2",
+    "expo-font": "^12.0.10",
+    "expo-image-picker": "~15.1.0",
     "expo-linking": "~6.3.1",
-    "expo-notifications": "^0.28.18",
+    "expo-notifications": "^0.28.19",
     "expo-secure-store": "~13.0.2",
-    "expo-splash-screen": "~0.27.4",
+    "expo-splash-screen": "~0.27.7",
     "expo-sqlite": "^14.0.6",
     "expo-status-bar": "~1.12.1",
     "expo-web-browser": "^13.0.3",
@@ -117,7 +118,7 @@
     "lexical": "^0.29.0",
     "markdown-it": "^14.1.0",
     "memoize-one": "^6.0.0",
-    "metro-react-native-babel-preset": "latest",
+    "metro-react-native-babel-preset": "^0.77.0",
     "nativewind": "^4.1.23",
     "npx": "^10.2.2",
     "path": "^0.12.7",
@@ -180,7 +181,5 @@
     "prettier": "^3.3.3",
     "typescript": "~5.3.3"
   },
-  "private": true,
-  "packageManager": "pnpm@8.15.9",
-  "pnpm": {}
+  "private": true
 }
